:zap: validation & improvement on repository

+++ lib/app/repository/local/local_data_source.dart
@@ -1,9 +1,9 @@
 /// This is interface for every app event
 /// Especially get data from local db
 abstract class LocalDataSource {
-  Future<void> setKiloPrice(int value);
+  Future<void> setKiloPrice(double value);
 
-  Future<int> getKiloPrice();
+  Future<double> getKiloPrice();
 
   /// Style & Theme
   Future<void> setIsDarkTheme(bool value);
